Split oversized link preview modules
