+++ app/api/wikimedia-card/route.ts
@@ -1,5 +1,4 @@
-import { toASCII } from "node:punycode";
-import type { NextRequest} from "next/server";
+import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
 
 import type {
@@ -8,13 +7,20 @@ import type {
   WikimediaImage,
 } from "@/services/api/wikimedia-card";
 
+import { assertPublicUrl, parsePublicUrl } from "@/lib/security/urlGuard";
 import {
-  UrlGuardError,
-  assertPublicUrl,
-  fetchPublicUrl,
-  parsePublicUrl,
-  type UrlGuardHostPolicy,
-} from "@/lib/security/urlGuard";
+  WIKIMEDIA_HOST_POLICY,
+  ensureWikimediaUrl,
+  fetchJson,
+  normalizeTarget,
+  parseAcceptLanguage,
+  respondWithGuardError,
+  sanitizeCacheKey,
+  type CommonsFileTarget,
+  type NormalizedTarget,
+  type SummaryTarget,
+  type WikidataTarget,
+} from "./target";
 import { sanitizeHtmlToText } from "@/lib/text/html";
 import LruTtlCache from "@/lib/cache/lruTtl";
 
@@ -35,512 +41,6 @@ const cache = new LruTtlCache<string, WikimediaCardResponse>({
   max: CACHE_MAX_ITEMS,
   ttlMs: TTL_BY_KIND.unavailable,
 });
-
-interface SummaryTarget {
-  readonly type: "summary";
-  readonly host: string;
-  readonly title: string;
-  readonly source: WikimediaSource;
-  readonly fragment?:
-    | { readonly raw: string; readonly display: string }
-    | undefined;
-  readonly canonicalFallback: string;
-}
-
-interface CommonsFileTarget {
-  readonly type: "commons-file";
-  readonly fileName: string;
-  readonly fragment?:
-    | { readonly raw: string; readonly display: string }
-    | undefined;
-}
-
-interface WikidataTarget {
-  readonly type: "wikidata";
-  readonly id: string;
-  readonly fragment?:
-    | { readonly raw: string; readonly display: string }
-    | undefined;
-}
-
-type NormalizedTarget = SummaryTarget | CommonsFileTarget | WikidataTarget;
-
-const decodeFragment = (
-  fragment: string
-): { raw: string; display: string } | undefined => {
-  if (!fragment) {
-    return undefined;
-  }
-
-  const raw = fragment;
-  let decoded = fragment;
-  try {
-    decoded = decodeURIComponent(fragment);
-  } catch {
-    decoded = fragment;
-  }
-
-  const display = decoded.replaceAll("_", " ").trim();
-  if (!display) {
-    return undefined;
-  }
-
-  return { raw, display };
-};
-
-const sanitizeCacheKey = (url: URL): string => {
-  const clone = new URL(url.toString());
-  clone.hash = "";
-  return clone.toString();
-};
-
-const parseAcceptLanguage = (header: string | null): string[] => {
-  if (!header) {
-    return ["en"];
-  }
-
-  const entries = header
-    .split(",")
-    .map((token) => {
-      const [langPart, qPart] = token.trim().split(";q=");
-      const lang = langPart?.toLowerCase();
-      const q = qPart ? Number.parseFloat(qPart) : 1;
-      return { lang, q: Number.isNaN(q) ? 1 : q };
-    })
-    .filter((entry) => entry.lang);
-
-  entries.sort((a, b) => b.q - a.q);
-
-  const preferred = new Set<string>();
-  for (const entry of entries) {
-    if (!entry.lang) {
-      continue;
-    }
-    preferred.add(entry.lang);
-    const base = entry.lang.split("-")[0];
-    if (base) {
-      preferred.add(base);
-    }
-  }
-
-  preferred.add("en");
-
-  return Array.from(preferred);
-};
-
-const respondWithGuardError = (
-  error: unknown,
-  fallbackMessage: string,
-  fallbackStatus = 400
-) => {
-  if (error instanceof UrlGuardError) {
-    return NextResponse.json(
-      { error: error.message },
-      { status: error.statusCode }
-    );
-  }
-
-  const message = error instanceof Error ? error.message : fallbackMessage;
-  return NextResponse.json({ error: message }, { status: fallbackStatus });
-};
-
-const WIKIMEDIA_ALLOWED_HOSTS = new Set([
-  "w.wiki",
-  "wikidata.org",
-  "www.wikidata.org",
-  "commons.wikimedia.org",
-  "upload.wikimedia.org",
-]);
-
-const WIKIMEDIA_HOST_POLICY: UrlGuardHostPolicy = {
-  allowedHosts: Array.from(WIKIMEDIA_ALLOWED_HOSTS),
-  allowedHostSuffixes: ["wikipedia.org"],
-};
-
-const isAllowedWikimediaHost = (hostname: string): boolean => {
-  if (WIKIMEDIA_ALLOWED_HOSTS.has(hostname)) {
-    return true;
-  }
-  return hostname.endsWith(".wikipedia.org");
-};
-
-const stripTrailingDots = (value: string): string => {
-  let end = value.length;
-  while (end > 0 && value.codePointAt(end - 1) === 46) {
-    end -= 1;
-  }
-  return end === value.length ? value : value.slice(0, end);
-};
-
-const ensureWikimediaUrl = (url: URL): void => {
-  let asciiHostname: string;
-  try {
-    asciiHostname = toASCII(url.hostname);
-  } catch {
-    throw new Error("Unsupported Wikimedia host");
-  }
-
-  const normalizedHostname = stripTrailingDots(asciiHostname).toLowerCase();
-  if (!normalizedHostname || !isAllowedWikimediaHost(normalizedHostname)) {
-    throw new Error("Unsupported Wikimedia host");
-  }
-};
-
-type WikimediaRequestInit = RequestInit & {
-  readonly timeoutMs?: number | undefined;
-  readonly language?: string | undefined;
-};
-
-const fetchWithTimeout = async (
-  input: string | URL,
-  init: WikimediaRequestInit = {}
-): Promise<Response> => {
-  let targetUrl: URL;
-  try {
-    targetUrl =
-      typeof input === "string"
-        ? parsePublicUrl(input)
-        : new URL(input.toString());
-  } catch (error) {
-    if (error instanceof UrlGuardError) {
-      throw error;
-    }
-    throw new UrlGuardError("Invalid URL provided.", "invalid-url", 400, {
-      cause: error,
-    });
-  }
-
-  ensureWikimediaUrl(targetUrl);
-  await assertPublicUrl(targetUrl, { policy: WIKIMEDIA_HOST_POLICY });
-
-  const { timeoutMs = REQUEST_TIMEOUT_MS, language, ...requestInit } = init;
-  const headers = new Headers(requestInit.headers);
-  if (!headers.has("accept")) {
-    headers.set("accept", "application/json");
-  }
-  if (language) {
-    headers.set("accept-language", language);
-  }
-
-  return fetchPublicUrl(
-    targetUrl,
-    {
-      ...requestInit,
-      headers,
-    },
-    {
-      timeoutMs,
-      userAgent: USER_AGENT,
-      maxRedirects: SHORT_LINK_MAX_REDIRECTS,
-      policy: WIKIMEDIA_HOST_POLICY,
-    }
-  );
-};
-
-const fetchJson = async <T>(
-  input: string | URL,
-  init: WikimediaRequestInit = {}
-): Promise<T> => {
-  const response = await fetchWithTimeout(input, init);
-  if (!response.ok) {
-    throw new UrlGuardError(
-      `Request failed with status ${response.status}`,
-      "fetch-failed",
-      response.status
-    );
-  }
-  return response.json() as Promise<T>;
-};
-
-const resolveShortLink = async (url: URL): Promise<URL> => {
-  ensureWikimediaUrl(url);
-  await assertPublicUrl(url, { policy: WIKIMEDIA_HOST_POLICY });
-
-  const response = await fetchPublicUrl(
-    url,
-    { method: "HEAD" },
-    {
-      timeoutMs: REQUEST_TIMEOUT_MS,
-      userAgent: USER_AGENT,
-      maxRedirects: SHORT_LINK_MAX_REDIRECTS,
-      policy: WIKIMEDIA_HOST_POLICY,
-    }
-  );
-
-  let finalResponse = response;
-
-  if (!response.ok) {
-    finalResponse = await fetchPublicUrl(
-      url,
-      {},
-      {
-        timeoutMs: REQUEST_TIMEOUT_MS,
-        userAgent: USER_AGENT,
-        maxRedirects: SHORT_LINK_MAX_REDIRECTS,
-        policy: WIKIMEDIA_HOST_POLICY,
-      }
-    );
-  }
-
-  if (!finalResponse.ok) {
-    throw new UrlGuardError(
-      "Unable to resolve short link",
-      "fetch-failed",
-      finalResponse.status
-    );
-  }
-
-  const finalUrl = finalResponse.url ? new URL(finalResponse.url) : url;
-  ensureWikimediaUrl(finalUrl);
-  await assertPublicUrl(finalUrl, { policy: WIKIMEDIA_HOST_POLICY });
-
-  return finalUrl;
-};
-
-const stripNamespace = (title: string): string => title.trim();
-
-const disallowedNamespaces = [
-  "special:",
-  "user:",
-  "talk:",
-  "user talk:",
-  "wikipedia talk:",
-  "file talk:",
-  "template:",
-  "template talk:",
-  "help:",
-  "help talk:",
-  "mediawiki:",
-  "mediawiki talk:",
-  "portal:",
-  "portal talk:",
-];
-
-const isDisallowedNamespace = (title: string): boolean => {
-  const lower = title.toLowerCase();
-  return disallowedNamespaces.some((ns) => lower.startsWith(ns));
-};
-
-const normalizeTitle = (title: string): string => {
-  const trimmed = title.trim();
-  if (!trimmed) {
-    return "";
-  }
-  return trimmed.replaceAll(/\s+/g, "_");
-};
-
-const resolveWikipediaTitle = async (
-  lang: string,
-  params: URLSearchParams
-): Promise<string | null> => {
-  if (params.has("title")) {
-    return params.get("title");
-  }
-
-  if (params.has("curid")) {
-    const pageId = params.get("curid");
-    if (!pageId) {
-      return null;
-    }
-    const query = new URLSearchParams({
-      action: "query",
-      format: "json",
-      pageids: pageId,
-      prop: "info",
-    });
-    const response = await fetchJson<{
-      readonly query?:
-        | {
-            readonly pages?:
-              | Record<string, { readonly title?: string | undefined }>
-              | undefined;
-          }
-        | undefined;
-    }>(`https://${lang}.wikipedia.org/w/api.php?${query.toString()}`);
-    const pages = response.query?.pages;
-    if (!pages) {
-      return null;
-    }
-    const page = Object.values(pages)[0];
-    return page?.title ?? null;
-  }
-
-  if (params.has("oldid")) {
-    const revisionId = params.get("oldid");
-    if (!revisionId) {
-      return null;
-    }
-    const query = new URLSearchParams({
-      action: "query",
-      format: "json",
-      prop: "revisions",
-      revids: revisionId,
-    });
-    const response = await fetchJson<{
-      readonly query?:
-        | {
-            readonly revisions?:
-              | Record<
-                  string,
-                  {
-                    readonly pageid?: number | undefined;
-                    readonly title?: string | undefined;
-                  }
-                >
-              | undefined;
-            readonly pages?:
-              | Record<string, { readonly title?: string | undefined }>
-              | undefined;
-          }
-        | undefined;
-    }>(`https://${lang}.wikipedia.org/w/api.php?${query.toString()}`);
-    const pages = response.query?.pages;
-    if (!pages) {
-      return null;
-    }
-    const page = Object.values(pages)[0];
-    return page?.title ?? null;
-  }
-
-  return null;
-};
-
-const extractCommonsFileName = (url: URL): string | null => {
-  const path = url.pathname;
-  if (!path) {
-    return null;
-  }
-
-  if (path.startsWith("/wiki/File:")) {
-    return decodeURIComponent(path.slice("/wiki/File:".length));
-  }
-
-  const segments = path.split("/").filter(Boolean);
-  if (segments.length === 0) {
-    return null;
-  }
-
-  if (segments[0] === "wikipedia" && segments[1] === "commons") {
-    if (segments[2] === "thumb" && segments.length >= 6) {
-      if (!segments[5]) {
-        return null;
-      }
-      return decodeURIComponent(segments[5]);
-    }
-    return decodeURIComponent(segments.at(-1)!);
-  }
-
-  return decodeURIComponent(segments.at(-1)!);
-};
-
-const normalizeTarget = async (url: URL): Promise<NormalizedTarget> => {
-  ensureWikimediaUrl(url);
-
-  let currentUrl = url;
-  if (currentUrl.hostname.toLowerCase() === "w.wiki") {
-    currentUrl = await resolveShortLink(currentUrl);
-    ensureWikimediaUrl(currentUrl);
-  }
-
-  let hostname = currentUrl.hostname.toLowerCase();
-  if (hostname.startsWith("www.")) {
-    hostname = hostname.slice(4);
-  }
-
-  const fragment = decodeFragment(
-    currentUrl.hash ? currentUrl.hash.slice(1) : ""
-  );
-
-  if (
-    hostname.endsWith(".wikipedia.org") ||
-    hostname.endsWith(".m.wikipedia.org")
-  ) {
-    let langHost = hostname;
-    if (langHost.endsWith(".m.wikipedia.org")) {
-      langHost = langHost.replace(".m.wikipedia.org", ".wikipedia.org");
-    }
-    if (langHost.startsWith("m.")) {
-      langHost = langHost.slice(2);
-    }
-    const lang = langHost.replace(".wikipedia.org", "");
-    const canonicalHost = `${lang}.wikipedia.org`;
-
-    let title: string | null = null;
-
-    if (currentUrl.pathname.startsWith("/wiki/")) {
-      const raw = currentUrl.pathname.slice(6);
-      title = decodeURIComponent(raw);
-    } else if (currentUrl.pathname.startsWith("/w/")) {
-      const params = currentUrl.searchParams;
-      title = await resolveWikipediaTitle(lang, params);
-    }
-
-    if (!title) {
-      throw new Error("Unable to resolve Wikipedia title");
-    }
-
-    const stripped = stripNamespace(title);
-    if (isDisallowedNamespace(stripped)) {
-      throw new Error("Namespace is not supported");
-    }
-
-    const normalizedTitle = normalizeTitle(stripped);
-    const canonicalFallback = `https://${canonicalHost}/wiki/${encodeURIComponent(
-      normalizedTitle
-    )}`;
-
-    return {
-      type: "summary",
-      host: canonicalHost,
-      title: normalizedTitle,
-      source: "wikipedia",
-      fragment,
-      canonicalFallback,
-    };
-  }
-
-  if (hostname === "commons.wikimedia.org") {
-    if (currentUrl.pathname.startsWith("/wiki/File:")) {
-      const fileName = currentUrl.pathname.slice("/wiki/File:".length);
-      return { type: "commons-file", fileName, fragment };
-    }
-
-    const titleSegment = currentUrl.pathname.startsWith("/wiki/")
-      ? currentUrl.pathname.slice(6)
-      : currentUrl.pathname.replace(/^\//, "");
-    const title = decodeURIComponent(titleSegment);
-    const normalizedTitle = normalizeTitle(title);
-    const canonicalFallback = `https://commons.wikimedia.org/wiki/${encodeURIComponent(
-      normalizedTitle
-    )}`;
-    return {
-      type: "summary",
-      host: "commons.wikimedia.org",
-      title: normalizedTitle,
-      source: "wikimedia-commons",
-      fragment,
-      canonicalFallback,
-    };
-  }
-
-  if (hostname === "upload.wikimedia.org") {
-    const fileName = extractCommonsFileName(currentUrl);
-    if (!fileName) {
-      throw new Error("Unable to determine file name");
-    }
-    return { type: "commons-file", fileName, fragment };
-  }
-
-  if (hostname === "wikidata.org" || hostname === "www.wikidata.org") {
-    const path = currentUrl.pathname.startsWith("/wiki/")
-      ? currentUrl.pathname.slice(6)
-      : currentUrl.pathname.replace(/^\//, "");
-    const id = path.toUpperCase();
-    return { type: "wikidata", id, fragment };
-  }
-
-  throw new Error("Unsupported Wikimedia host");
-};
 
 const sanitizeHtml = (value: string): string => {
   return sanitizeHtmlToText(value, { preserveTagSpacing: true })
@@ -651,8 +151,8 @@ const buildSummaryCard = async (
           description: page.description
             ? sanitizeHtml(page.description)
             : page.extract
-            ? sanitizeHtml(page.extract)
-            : null,
+              ? sanitizeHtml(page.extract)
+              : null,
           url: page.content_urls?.desktop?.page ?? "",
           thumbnail: page.thumbnail?.source
             ? {
@@ -806,13 +306,13 @@ const buildCommonsCard = async (
           alt: description ?? page.title ?? fileTitle,
         }
       : info.url
-      ? {
-          url: info.url,
-          width: info.width,
-          height: info.height,
-          alt: description ?? page.title ?? fileTitle,
-        }
-      : null;
+        ? {
+            url: info.url,
+            width: info.width,
+            height: info.height,
+            alt: description ?? page.title ?? fileTitle,
+          }
+        : null;
 
     const original:
       | (WikimediaImage & { readonly mime?: string | null | undefined })
