+++ scripts/debt-ratchet-baseline.json
@@ -4,22 +4,13 @@
   "counts": {
     "any_casts": 0,
     "todo_comments": 0,
-    "oversized_files": 36,
+    "oversized_files": 24,
     "legacy_wordpress_runtime": 0,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
   },
   "oversized_file_allowlist": [
-    "app/api/farcaster/route.ts",
-    "app/api/open-graph/6529/service.ts",
-    "app/api/open-graph/compound/service.ts",
-    "app/api/open-graph/opensea/service.ts",
-    "app/api/open-graph/route.ts",
-    "app/api/open-graph/transient/service.ts",
-    "app/api/open-graph/utils.ts",
-    "app/api/pepe/resolve/route.ts",
-    "app/api/wikimedia-card/route.ts",
     "components/about/AboutTermsOfService.tsx",
     "components/auth/SeizeConnectProvider.tsx",
     "components/brain/left-sidebar/waves/memes-quick-vote/useMemesQuickVotePreviewSwipe.ts",
@@ -37,15 +28,12 @@
     "components/react-query-wrapper/ReactQueryWrapper.tsx",
     "components/rememes/RememePage.tsx",
     "components/user/collected/UserPageCollected.tsx",
-    "components/waves/TwitterPreviewCard.tsx",
     "components/waves/WaveTrustSignals.tsx",
     "components/waves/drops/WaveDrop.tsx",
     "components/waves/drops/WaveDropReactions.tsx",
-    "components/waves/github/GithubLinkPreview.tsx",
     "contexts/wave/hooks/useWaveRealtimeUpdater.ts",
     "helpers/Helpers.ts",
     "hooks/useMemesQuickVoteQueue.ts",
-    "lib/twitter/fetcher.ts",
     "services/api/wave-drops-v2-api.ts"
   ]
 }
